app-wse/pyproject.toml
[tool.poetry]
name = "wselfedu"
version = "0.1.0"
description = ""
authors = ["Sergei Mikurov <454004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
django = "4.2"
django-bootstrap5 = "^24.2"
django-filter = "^24.2"
django-crispy-forms = "^2.1"
django-jsonview = "^2.0.0"
crispy-bootstrap5 = "^2024.2"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.0"
django-debug-toolbar = "^4.3.0"
pytest-django = "^4.8.0"
requests = "^2.32.2"
pytest-cov = "^5.0.0"
flake8-docstrings = "^1.7.0"
sphinx = "^7.3.7"
sphinx-rtd-theme = "^2.0.0"
myst-parser = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage]
    [tool.coverage.run]
    omit = [
        'tests/*',
        '*/tests/*',
        '*/migrations/*',
        '*/apps.py',
        'manage.py',
        'docs/*',
    ]

    [tool.coverage.report]
    skip_empty = true